Fix bind service failure on Android Auto
The failure is because ActiveServices thought window context's token is
an Activity token but could not find the corresponding ActivityRecord
in ATMS and then early-returned. This issue was exposed because of
Presentation's window context migration.
This CL relaxed the condition to early return only if the token is
neither an Activity token nor a WindowContext token.

Bug: 171280916
Bug: 171027173
Bug: 170960206

Test: Tests mentioned in b/170960206#comment1
Test: manual - use auto desktop mode to launch Music/Maps
Test: atest WindowContextTests#testWindowContextBindService

Change-Id: I313d3d870caa28e02af035b4a6032d649f81a510
(cherry picked from commit b6eda4c8ee99c160c4de8c8018b55ac3ae6c92d0)