Use QSPanelController in QSAnimator.
The animator doesn't want the view, but rather some of the views
contained within. At times, it simply wants some of the functionality
that the view provides that will be better encapsulated by its
controller.

Replaced Dependency.get(TunerService.class) with an injected
instance.

Injected an executor for the class to use rather than relying on
QSPanels implicit handler.

Bug: 168904199
Test: manual
Change-Id: Id228a9dd921fbb11ed2d1d56c93d37d8ceda1cd3