Synchronize destination bounds after rotation in MotionHelper
Make sure that the bounds state is updated with the final
destination bounds after a rotation *before* sending the
pip animation finished & cancelled callbacks.
This ensures that the MotionHelper bounds are synchronized
correctly at the end of the rotation to the post-rotation
bounds and not the pre-rotation bounds by
PipTouchHandler#onPinnedStackAnimationEnded.

Bug: 171352429
Test: Enter PIP, rotate, move PIP - the window doesn't jump
to the pre-rotation bounds when trying to move it.

Change-Id: I36eb192210ec58ac3448c423f43b959ce29eff1f