Reland "Distinguish warm/hot launch by whether the activity has attached process"
If the process of an activity was died in background, its activity and
task record will still be in history stack. If the process is started
again in background such as receiving broadcast. Then the launch of the
same activity will have the conditions: process exists and the result
code is START_TASK_TO_FRONT. But the activity still need to create so
it should be warm launch.

This reverts commit 1de14357fe90273510fa3041a31fa8cb5f0dd022.
Reason for revert: The original patch is no problem

Bug: 171549711
Test: ActivityMetricsLaunchObserverTests#testOnActivityLaunchFinished
Change-Id: I2b8e45ddfd76ce4fb0fdb0c244a8f4170fbabd21