Manual revert of 5s FGS notification enforcement
It had been policy to ensure that any notification representing a
foreground service be shown for a minimum of 5 seconds. This policy is
now going away.

In order to implement the policy before, we had to:
    1) ensure that any call to `notify()` eventually made it to any
    NotificationListenerService, regardless of a cancel that might
    happen
    2) Implement the policy in SystemUI

This CL is a manual revert of both behaviors. Manual because there had
been interim fixes along the way that make a mechanical revert much
messier.

Bug: 171513045
Test: atest SystemUITests
Change-Id: I2d9305dde80cf7b03f7096aaf1f81f7f53c49453