Split PermissionSettings into LegacyPermissionSettings and PermissionRegistry.
And split BasePermission into LegacyPermission and Permission.

This way permission definitions become independent from package
manager settings which enables using its own locking. The legacy
classes are only for reading, compat dumping, and temporary writing
before we migrate persistence.

Similar to how we handled permission state, we are not migrating the
persistence in this change, but writing back to the legacy class for
persistence.

The BasePermission reference is removed in LegacyPermissionState, as
having only the permission name is already sufficient.

The "requested permissions" section in dump no longer prints whether a
the requested permission is restricted - the info is still available
in the permissions section.

Bug: 168319670
Test: manual
Change-Id: Ieac5f4005f6f04dde4a88592dec534f31f298624