Update handling of incoming CEC power key presses
Update handling of <User Control Pressed> messages with power related
operands. Power commands change the power status properly and should not
be converted to power keycodes.

This particularly changes the handling of <User Control Pressed>["Power"]
which could be converted to KEYCODE_POWER before when the device was on
when the keycode was received, causing the device to turn off. This is
no longer supported and <User Control Pressed>["Power"] will be treated
as a power on command.

Test: atest HdmiCecLocalDeviceTest
Fixes: 174452275
Change-Id: I10f6d0d235d999bd29b54ba96728c0bda9697ec5