Reduce severity of logging statement
It's often not an issue, much less a major or severe issue, to
not have some of the adb keys on a device.  So we reduce the
severity of the logging in this case.

We especially care about this instance because dump() is invoked
when generating a bugreport, when folks are likely to be looking
through for Errors.  Giving an Error for this non-Error situation
can confuse triage of other issues.

Test: Bug report on device lacking user adb keys no longer has an Error about it
Bug: 174797634
Change-Id: Ib17efc2e7c61ff5e0bbe00a37b812fd198c2b5fa