Mark layout_preference container not important for a11y.
This container by itself does not have anything meaningful for
accessibility. Marking this view not important for a11y forces tools
such as talkback look into children views for accessibility action.

However if this view is left at default accessibility importantace, it
tries to aggregate accessibilty info from children. So if there is any
clickable children, this view will say "double tap to activate" when
it's not a tap target.

Fix: 175515538
Fix: 175515122
Test: manual
Change-Id: I415090fb8601bc4cb129999a50ac35e045ef7830