Ensure OK/Cancel buttons always shown
There is some weirdness with the way the AlertActivity behaves when
it's populated and elements are enabled while it's not actually on
the screen. Work around this by populating the AlertActivity during
onResume() [eg. every time the activity is shown] instead of
onCreate() [eg. only the first time the activity is created].

Fixes: 172699073, 173159446
Test: Manual
Test: Start a staging install
Test: Turn off the phone screen
Test: Wait for the staging to complete
Test: Turn on the phone screen.
Test: See the install dialog has OK/Cancel buttons
Change-Id: Ia9525f20f98f42e6b2b865fa926b9707d416cde4